+++ lib/garuda.ex
@@ -1,5 +1,3 @@
 defmodule Garuda do
-  @moduledoc """
-   This is the documentation for the Garuda Project.
-  """
+  @moduledoc false
 end