+++ lib/matchmaker/matchmaker_channel.ex
@@ -4,7 +4,12 @@ defmodule Garuda.MatchMakerChannel do
   alias Garuda.MatchMaker.Matcher
 
   def join("garuda_matchmaker:lobby", match_details, socket) do
-    match_id = Matcher.join_or_create(match_details)
-    {:ok, %{"match_id" => match_id}, socket}
+    match_resp = Matcher.join_or_create(match_details)
+
+    if Map.has_key?(match_resp, "match_id") do
+      {:ok, match_resp, socket}
+    else
+      {:error, match_resp}
+    end
   end
 end