+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.1
+  - Fault-tolerant RoomDb.
+  - Adding current players in a room in orwell (monitor dashboard).
+  - Handling MatchId exists case in matchmaker.
+  - Guides for using Garuda in a project.
+## 0.2.0
+  - Documentation.
+  - Refactoring.
+  - Deploying the demo.
 ## 0.2.0-rc.3
   - Orwell restructured like live_dashboard.
   - Orwell Bug fixes + Refactoring